+++ CHANGELOG.md
@@ -7,6 +7,7 @@ VERSION 4.0.x-dev
 1. Removed renderPageWrapper method from CRUDHandler.
 2. Renamed `findTemplatePath` with `findTemplate`.
 3. Renamed `getDefaultData` with `getDefaultRecordArgs`.
+4. `loadRecord` is now `loadCurrentRecord` and only returns record if loaded.
 
 
 VERSION 3.0.x-dev